Finish SDK workflow and restore architecture budgets
