+++ crates/connect-agent/src/server/operations.rs
@@ -1,90 +1,5 @@
 use super::{metrics, operation_responses::*, *};
 impl AgentState {
-    fn scoped_operation(
-        &self,
-        transport: &'static str,
-        collection_id: uuid::Uuid,
-        operation: &str,
-        input: &serde_json::Value,
-        grant: &mdbase_connect_protocol::GrantSummary,
-    ) -> Result<serde_json::Value, ConnectError> {
-        let started = Instant::now();
-        let synchronize_us = std::cell::Cell::new(0_u64);
-        if matches!(
-            operation,
-            "assess_collection_setup" | "apply_collection_setup"
-        ) {
-            validate_collection_setup_binding(input, grant)?;
-        }
-        if matches!(
-            operation,
-            "list_timers" | "put_timer" | "cancel_timer" | "reconcile_timers"
-        ) {
-            let result = self
-                .runtime_timers
-                .as_ref()
-                .ok_or_else(|| {
-                    ConnectError::TimerRuntime("The timer authority is unavailable.".to_string())
-                })
-                .and_then(|timers| {
-                    timers
-                        .operation(collection_id, grant.clone(), operation, input.clone())
-                        .map_err(|error| {
-                            if error.internal {
-                                ConnectError::TimerRuntime(error.message)
-                            } else {
-                                ConnectError::InvalidTimer(error.message)
-                            }
-                        })
-                });
-            profile_operation(transport, operation, started, 0, &result);
-            return result;
-        }
-        if operation == "sync" {
-            let mode = if grant.operations.iter().any(|operation| {
-                matches!(
-                    operation.as_str(),
-                    "create" | "update" | "delete" | "rename"
-                )
-            }) {
-                SyncReplicaMode::ReadWrite
-            } else {
-                SyncReplicaMode::ReadOnly
-            };
-            let result = self.registry.sync_operation_synchronized(
-                collection_id,
-                input,
-                LocalReplica {
-                    id: grant.id,
-                    name: grant.application_name.clone(),
-                    mode,
-                    allowed_types: Default::default(),
-                },
-                &grant.scope,
-                |invalidation| {
-                    let synchronize_started = Instant::now();
-                    self.watcher.synchronize(collection_id, invalidation);
-                    synchronize_us.set(elapsed_us(synchronize_started));
-                },
-            );
-            profile_operation(transport, operation, started, synchronize_us.get(), &result);
-            return result;
-        }
-        let result = self.registry.scoped_operation_synchronized(
-            collection_id,
-            operation,
-            input,
-            &grant.scope,
-            |invalidation| {
-                let synchronize_started = Instant::now();
-                self.watcher.synchronize(collection_id, invalidation);
-                synchronize_us.set(elapsed_us(synchronize_started));
-            },
-        );
-        profile_operation(transport, operation, started, synchronize_us.get(), &result);
-        result
-    }
-
     pub(super) fn local_operation(
         &self,
         collection_id: uuid::Uuid,
@@ -1004,27 +919,4 @@ impl AgentState {
         }
         message
     }
-}
-
-fn validate_collection_setup_binding(
-    input: &serde_json::Value,
-    grant: &mdbase_connect_protocol::GrantSummary,
-) -> Result<(), ConnectError> {
-    let setup = if input.get("setup").is_some() {
-        &input["setup"]
-    } else {
-        input
-    };
-    let application_id = setup["application_id"].as_str();
-    let declaration_digest = setup["declaration_digest"].as_str();
-    if application_id != Some(grant.application_declaration_id.as_str())
-        || declaration_digest
-            != Some(format!("sha256:{}", grant.application_manifest_digest).as_str())
-    {
-        return Err(ConnectError::AccessDenied(
-            "Collection setup must match the exact application declaration bound to this grant."
-                .to_string(),
-        ));
-    }
-    Ok(())
 }
